% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd.R
\name{copy_markdown_report}
\alias{copy_markdown_report}
\title{Copy the reporting R Markdown file to a chosen path}
\usage{
copy_markdown_report(path)
}
\arguments{
\item{the}{path to copy to.}
}
\value{
The location of the R markdown file included in the package (for copying etc)
}
\description{
Copy the reporting R Markdown file to a chosen path
}
