% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cos_guidelines.R
\name{cos_grade_guidelines}
\alias{cos_grade_guidelines}
\title{College of Sciences guidelines for the percentage of passing students to be awarded a grade category}
\usage{
cos_grade_guidelines(level)
}
\arguments{
\item{level}{the level (100, 200, 300, 700) to retrieve guidelines for}
}
\value{
A tibble with columns `gradecat`, `min`, `max`
}
\description{
College of Sciences guidelines for the percentage of passing students to be awarded a grade category
}
