% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cos_guidelines.R
\name{cos_pass_guidelines}
\alias{cos_pass_guidelines}
\title{College of Science guidelines for the percentage of graded students who should pass a course.}
\usage{
cos_pass_guidelines(level)
}
\arguments{
\item{level}{the level (100, 200, 300, 700) to retrieve guidelines for}
}
\value{
A one-row tibble with columns `min`, `max`
}
\description{
College of Science guidelines for the percentage of graded students who should pass a course.
}
