% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sms_reporting.R
\name{download_from_sms}
\alias{download_from_sms}
\title{Download grades from SMS reporting}
\usage{
download_from_sms(paper, year, semester, location = "ALL", credentials)
}
\arguments{
\item{paper}{the paper to download}

\item{year}{the year to download}

\item{semester}{the semester to download}

\item{location}{the location to download, defaults to "ALL". Other values are "MTUI", "AKLI", "DISD"}

\item{credentials}{the credentials from `get_credentials`.}
}
\value{
A tibble with columns `paper`, `mode`, `year`, `semester`,
`number`, `type`, `weight`, `student_id`
`surname`, `firstname`, `mark`, `grade`, `agreed_mark`, `agreed_grade`,
`overall_mark`, `overall_grade`, `agreed_overall_mark`, `agreed_overall_grade`,
`minutes`.
}
\description{
Download grades from SMS reporting
}
\seealso{
`read_sms_reporting`, `parse_sms_reporting`
}
