% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sms_reporting.R
\name{download_raw_from_sms}
\alias{download_raw_from_sms}
\title{Download raw grades from SMS reporting}
\usage{
download_raw_from_sms(paper, year, semester, location = "ALL", credentials)
}
\arguments{
\item{paper}{the paper to download}

\item{year}{the year to download}

\item{semester}{the semester to download}

\item{location}{the location to download, defaults to "ALL". Other values are "MTUI", "AKLI", "DISD"}

\item{credentials}{the credentials from `get_credentials`.}
}
\value{
The raw marks.
}
\description{
Download raw grades from SMS reporting
}
\seealso{
`read_sms_reporting`, `parse_sms_reporting`
}
