% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sms_report.R
\name{extract_paper_information}
\alias{extract_paper_information}
\title{Extract paper information from SMS data read using `read_sms_reporting_csv()`}
\usage{
extract_paper_information(data)
}
\arguments{
\item{data}{the data read via `read_sms_reporting_csv()`.}
}
\value{
A list with entries `paper`, `year`, `semester`, `level`, `students`.
}
\description{
Extract paper information from SMS data read using `read_sms_reporting_csv()`
}
\seealso{
`read_sms_reporting_csv`
}
