% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sms_reporting.R
\name{get_credentials}
\alias{get_credentials}
\title{Authenticate credentials on the Massey system}
\usage{
get_credentials(user = NULL, pass = NULL)
}
\arguments{
\item{user}{an (optional) user name, defaults to NULL}

\item{pass}{an (optional) password, defaults to NULL}
}
\value{
A `list` with values `user` and `pass`
}
\description{
Authenticate credentials on the Massey system
}
