% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd.R
\name{get_markdown_report}
\alias{get_markdown_report}
\title{Get the reporting R Markdown file location}
\usage{
get_markdown_report()
}
\value{
The location of the R markdown file included in the package (for copying etc)
}
\description{
Get the reporting R Markdown file location
}
