% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sms_reporting.R
\name{get_url_for_sms}
\alias{get_url_for_sms}
\title{Get URL for SMS reporting for a paper}
\usage{
get_url_for_sms(paper, year, semester, location = "ALL")
}
\arguments{
\item{paper}{the paper to download}

\item{year}{the year to download}

\item{semester}{the semester to download}

\item{location}{the location to download, defaults to "ALL". Other values are "MTUI", "AKLI", "DISD"}
}
\value{
The URL.
}
\description{
Get URL for SMS reporting for a paper
}
\seealso{
`read_sms_reporting`, `parse_sms_reporting`
}
