% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{grade_category}
\alias{grade_category}
\title{Computes the grade category (A, B, C, F, DC, NF) for a vector of grades}
\usage{
grade_category(grades)
}
\arguments{
\item{grades}{the grades as a character vector}
}
\value{
a factor variable with grade categories
}
\description{
Computes the grade category (A, B, C, F, DC, NF) for a vector of grades
}
