% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{grade_factor}
\alias{grade_factor}
\title{Converts a grade character string to a factor}
\usage{
grade_factor(grades)
}
\arguments{
\item{grades}{the grades as a character vector}
}
\value{
a factor variable ordered appropriately
}
\description{
Converts a grade character string to a factor
}
