% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mark_imputation.R
\name{impute_all_marks}
\alias{impute_all_marks}
\title{Imputation of all marks via leave one out prediction of each student
mark.}
\usage{
impute_all_marks(marks)
}
\arguments{
\item{marks}{A tibble of marks with columns `student_id` and then the remaining assessment items in separate columns.}
}
\value{
A tibble with columns `student_id`, `assessement`, `mark`, `predicted`, `P`
}
\description{
Uses a simple linear model trained on all other students to infer a mark
and prediction interval for each assessment item for each student.
}
\details{
The prediction interval is then used to compute a P-value for each assessment
item, and these are corrected to control for false discovery rate.
}
