% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sms_report.R
\name{parse_raw_sms_data}
\alias{parse_raw_sms_data}
\title{Prase raw data from SMS reporting}
\usage{
parse_raw_sms_data(raw_sms)
}
\arguments{
\item{raw_sms}{a data.frame of data from SMS}
}
\value{
A tibble with columns `paper`, `mode`, `year`, `semester`,
`number`, `type`, `weight`, `student_id`
`surname`, `firstname`, `mark`, `grade`, `agreed_mark`, `agreed_grade`,
`overall_mark`, `overall_grade`, `agreed_overall_mark`, `agreed_overall_grade`,
`minutes`.
}
\description{
All grades are converted to factors.
}
\seealso{
`read_sms_reporting_csv`
}
