% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sms_report.R
\name{read_sms_reporting_csv}
\alias{read_sms_reporting_csv}
\title{Read .csv file from SMS reporting}
\usage{
read_sms_reporting_csv(path)
}
\arguments{
\item{path}{the path to the .csv file.}
}
\value{
A tibble with columns `paper`, `mode`, `year`, `semester`,
`number`, `type`, `weight`, `student_id`
`surname`, `firstname`, `mark`, `grade`, `agreed_mark`, `agreed_grade`,
`overall_mark`, `overall_grade`, `agreed_overall_mark`, `agreed_overall_grade`,
`minutes`.
}
\description{
All grades are converted to factors.
}
\seealso{
`extract_paper_information`
}
