% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{round_mark}
\alias{round_mark}
\title{Round a mark to the nearest integer}
\usage{
round_mark(marks)
}
\arguments{
\item{marks}{A vector of marks to round}
}
\value{
The rounded marks
}
\description{
The standard `round()` function rounds to nearest even. This routine
rounds to nearest integer regardless of even or odd (i.e. should be consistent)
with SMS
}
